% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compileAeth.R
\name{compileAeth}
\alias{compileAeth}
\title{Compile AE33 data files}
\usage{
compileAeth(path = "./aeth", ...)
}
\arguments{
\item{path}{The path to the folder containing the files to aggregate.}

\item{...}{Other options passed to \code{\link[utils]{read.table}}.}
}
\value{
Nothing. Side-effect: Writes compiled data to TSV.
}
\description{
Compile AE33 data files into a single file.
}
