% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBg.waves.R
\name{getBg.waves}
\alias{getBg.waves}
\title{Get Background Values for a Single Wavelength}
\usage{
getBg.waves(x, limits = c(300, 350), period = 1800)
}
\arguments{
\item{x}{A vector of data.}

\item{limits}{The limits in seconds to use for the background period of each run.}

\item{period}{The sample period (in seconds) including the background.}
}
\value{
A vector of background values.
}
\description{
Get background values for a single wavelength
}
\examples{
apply(X[, cols], 2, getBg.waves, ...) \\\%>\\\%
fillToLength() \\\%>\\\%
setNames(names)
}
\author{
Al Fischer
}
