% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTau0.R
\name{getTau0}
\alias{getTau0}
\title{Extract tau_0 From MultiPAS Data}
\usage{
getTau0(x, period = 1800)
}
\arguments{
\item{x}{A vector of cavity ringdown data.}

\item{period}{The sample period (in seconds) including the background.}
}
\value{
A vector of t_0 values.
}
\description{
Extracts tau_0 from a MultiPAS data frame for a cavity ringdown instrument.
}
