% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importAeth.R
\name{importAeth}
\alias{importAeth}
\title{Import AE33 Data}
\usage{
importAeth(path = "aeth/aethCompiled.txt", ...)
}
\arguments{
\item{path}{The path to the folder containing the file to import}

\item{...}{Other options passed to \code{\link[utils]{read.table}}.}
}
\value{
A data frame containing the file read.
}
\description{
Imports compiled AE33 data file output by \code{\link[aeRo]{compileAeth}}.
}
