% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ssa.R
\name{plot.ssa}
\alias{plot.ssa}
\title{Plotting function for an SSA Object}
\usage{
\method{plot}{ssa}(x, type = "l", ylim = c(0, 1), xlab = "time",
  xaxs = "i", ylab = "SSA", yaxs = "i", col = "red", lwd = 2,
  lty = "solid", axes = F, mar = c(5, 5, 2, 2), ...)
}
\arguments{
\item{x}{a list of class \code{ssa} with the elements \code{x} and \code{y}}

\item{...}{Other options passed to plot().}
}
\value{
Nothing. Side-effect: plots graphs.
}
\description{
Plots single scattering albedo vs. time given a list of SSA data.
}
