% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processSpectrum.R
\name{processSpectrum}
\alias{processSpectrum}
\title{Import and Plot MultiPAS Absorption/Extinction Data}
\usage{
processSpectrum(n, path = "./spectra", m)
}
\arguments{
\item{n}{a numeric indicating the sample number to process.}

\item{path}{the path to the data files.}
}
\value{
A data frame containing the processed spectrum
}
\description{
Calculate absorption and extinction from sample and background files.
}
